requirements.txt
pytesseract>=0.3.10
Pillow>=10.0.0
pdfplumber>=0.10.0
python-docx>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
google-cloud-vision>=3.7.0
google-cloud-storage>=2.16.0
firebase-admin>=6.5.0
